avoid duplicate code between windows/linux
